+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   RadioButton: fix helper text accessibility
+-   Switch: fix helper text accessibility
 
 ## [3.0.4][] - 2022-11-07